feature: expand scope of extend picture to cover all aspect ratios
